translations: Align bf|em mapping with README                                                                        | In commit 3bfdd040d657dcbf84c5f0a4ddfa9744a9f30816, the intention was to swap the "Reversed Hand With Middle Finger Extended" emoji with the "Up Pointing Backhand Index" emoji. The change to the README in that commit reflects that intention, and defines the thumbs-up/thumbs-down as opposing characters, and the point-up/point-down as opposing chars.
However, in that commit, the corresponding change to translations.py
changed the behavior differently. Instead of replacing the middle
finger emoji, that commit replaced the thumbs-up emoji. As a result,
Emojy code today still uses the middle-finger emoji in opposition with
the point-down emoji, and opposes the point-up emoji in opposition with
the thumbs-down emoji. You can verify that this is the current Emojy
syntax by inspecting the generated examples/*.em files.

This change updates translations.py in accordance with the original
intentions of 3bfdd040d657dcbf84c5f0a4ddfa9744a9f30816, and includes
a corresponding regeneration of examples/*.em.